Fix go-client / Refactor k8s util code
